Fix longstanding bug in reloading of task-form provider.
When doing a transition on a story, we now refresh the 'add task'
part.  We now have an unexpanded version of this part.  When clicking
on this, kss still makes sure like always to expand it by getting an
expanded version from the server.

About four years ago I fixed this about three times and it got broken
again so I gave up.  Now I am happy. :-)